backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
distlib==0.3.7
Django==4.2.2
django-cors-headers==4.2.0
django-location-field==2.7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
filelock==3.12.2
inflection==0.5.1
packaging==23.1
Pillow==10.0.0
pipenv==2023.8.23
platformdirs==3.10.0
psycopg==3.1.10
psycopg2-binary==2.9.7
PyJWT==2.7.0
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
uuid==1.30
virtualenv==20.24.3
